wasm/jsapi/esm-integration/memory-from-js.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script type=module>
setup({ single_test: true });
import {getFirstByte} from "./memory-from-js.wasm";
import {buffer} from "./memory-from-js.mjs";
assert_equals(getFirstByte(), buffer.buffer[0]);
done();
</script>
